extract data-utils.ts
